requirements.txt
openai==1.40.6
python-dotenv==1.0.1
gTTS==2.5.1
moviepy==1.0.3
imageio==2.34.1
imageio-ffmpeg==0.4.9
numpy==1.26.4
Pillow==10.3.0
decorator==4.4.2
proglog==0.1.10
requests==2.32.3
azure-storage-blob==12.22.0